Moved Batman to modules.exports instead of exports.Batman, revamped tools to work and look nice.
 - moved Batman to be the exported object instead of a property on it.
 - rewrote the tools to use cli and its handy stuff like auto generation of usage's
 - made the server compile CoffeeScript on the fly when asked for a nonexistent .js file in a directory with a .coffee file of the same name
 - made the server serve lib from the batman project root to routes matching /batman
 - updated the batman g generated files to work and be a bit more helpful with some boilerplate comments

+++ tests/batman/model_test.coffee
@@ -159,19 +159,20 @@ asyncTest "async", 2, ->
 
 QUnit.module "Batman.Model: storage"
 
-asyncTest "local storage", 1, ->
-  localStorage.clear()
-  
-  class Product extends Batman.Model
-    @persist Batman.LocalStorage
-    @encode 'foo'
-  
-  p = new Product foo: 'bar'
-  
-  p.afterSave ->
-    copy = Product.find p.id
-    copy.afterLoad ->
-      equal copy.get('foo'), 'bar'
-      start()
-  
-  p.save()
+if window? && 'localStorage' in window
+  asyncTest "local storage", 1, ->
+    localStorage.clear()
+    
+    class Product extends Batman.Model
+      @persist Batman.LocalStorage
+      @encode 'foo'
+    
+    p = new Product foo: 'bar'
+    
+    p.afterSave ->
+      copy = Product.find p.id
+      copy.afterLoad ->
+        equal copy.get('foo'), 'bar'
+        start()
+    
+    p.save()
